GAS-1160: Moved calibration to a separate subdirectory.
